+++ packages/sdk/src/lib/types/commonTypes.ts
@@ -117,3 +117,12 @@ export interface SearchableDataSchema
     | SearchableSchemaEntryType
     | SearchableSchemaEntryType[]
   > {}
+
+export type MatchOptions = {
+  useVoucher: boolean;
+  voucherAddress?: string;
+};
+
+export type DefaultWorkerpoolConsumer = {
+  defaultWorkerpool: AddressOrENS;
+};