+++ packages/sharing-smart-contract/contracts/DataProtectorSharing.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /******************************************************************************
- * Copyright 2024 IEXEC BLOCKCHAIN TECH                                       *
+ * Copyright 2024-2025 IEXEC BLOCKCHAIN TECH                                       *
  *                                                                            *
  * Licensed under the Apache License, Version 2.0 (the "License");            *
  * you may not use this file except in compliance with the License.           *
@@ -25,7 +25,7 @@ import {ERC721Holder} from "@openzeppelin/contracts/token/ERC721/utils/ERC721Hol
 import {Math} from "@openzeppelin/contracts/utils/math/Math.sol";
 import {IDataProtectorSharing, IexecLibOrders_v5, ICollection, ISubscription, IRental, ISale} from "./interfaces/IDataProtectorSharing.sol";
 import {AddOnlyAppWhitelistRegistry, IAddOnlyAppWhitelist} from "./registry/AddOnlyAppWhitelistRegistry.sol";
-import {ManageOrders, IExecPocoDelegate} from "./ManageOrders.sol";
+import {ManageOrders} from "./ManageOrders.sol";
 import {IRegistry} from "./interfaces/IRegistry.sol";
 
 /// @custom:oz-upgrades-unsafe-allow state-variable-immutable
@@ -57,7 +57,7 @@ contract DataProtectorSharing is
      **************************************************************************/
     /// @custom:oz-upgrades-unsafe-allow constructor
     constructor(
-        IExecPocoDelegate _proxy,
+        address _proxy,
         IRegistry protectedDataRegistry_,
         AddOnlyAppWhitelistRegistry addOnlyAppWhitelistRegistry_
     ) ManageOrders(_proxy) {
@@ -72,7 +72,7 @@ contract DataProtectorSharing is
         __AccessControl_init();
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
 
-        updateEnv("ipfs", "https://result.v8-bellecour.iex.ec");
+        updateEnv("ipfs");
     }
 
     /***************************************************************************
@@ -166,7 +166,12 @@ contract DataProtectorSharing is
             _workerpoolOrder.category
         );
 
-        dealid = POCO_DELEGATE.matchOrders(_appOrder, _datasetOrder, _workerpoolOrder, requestOrder);
+        dealid = POCO_DELEGATE.matchOrders(
+            _appOrder,
+            _datasetOrder,
+            _workerpoolOrder,
+            requestOrder
+        );
         emit ProtectedDataConsumed(dealid, _protectedData, _mode);
     }
 
@@ -178,7 +183,11 @@ contract DataProtectorSharing is
 
     // Overide burn function from ERC721BurnableUpgradeable.
     // Enable to burn a collectionTokenID.
-    function _update(address to, uint256 _collectionTokenId, address auth) internal virtual override returns (address) {
+    function _update(
+        address to,
+        uint256 _collectionTokenId,
+        address auth
+    ) internal virtual override returns (address) {
         CollectionDetails storage _collectionDetails = collectionDetails[_collectionTokenId];
         if (to == address(0)) {
             if (_collectionDetails.size > 0) {
@@ -192,7 +201,10 @@ contract DataProtectorSharing is
     }
 
     /// @inheritdoc IDataProtectorSharing
-    function getProtectedDataRenter(address _protectedData, address _renterAddress) public view returns (uint48) {
+    function getProtectedDataRenter(
+        address _protectedData,
+        address _renterAddress
+    ) public view returns (uint48) {
         return protectedDataDetails[_protectedData].renters[_renterAddress];
     }
 
@@ -205,7 +217,12 @@ contract DataProtectorSharing is
     }
 
     /// @inheritdoc IDataProtectorSharing
-    function receiveApproval(address _sender, uint256, address, bytes calldata _extraData) public returns (bool) {
+    function receiveApproval(
+        address _sender,
+        uint256,
+        address,
+        bytes calldata _extraData
+    ) public returns (bool) {
         if (msg.sender != address(POCO_DELEGATE)) {
             revert OnlyPocoCallerAuthorized(msg.sender);
         }
@@ -229,7 +246,10 @@ contract DataProtectorSharing is
             _rentProtectedData(protectedData, _sender, rentingParams);
             return true;
         } else if (selector == this.buyProtectedData.selector) {
-            (address protectedData, address to, uint72 price) = abi.decode(_extraData[4:], (address, address, uint72));
+            (address protectedData, address to, uint72 price) = abi.decode(
+                _extraData[4:],
+                (address, address, uint72)
+            );
             _buyProtectedData(protectedData, _sender, to, price);
             return true;
         }
@@ -241,11 +261,9 @@ contract DataProtectorSharing is
      *                         Admin                                           *
      **************************************************************************/
     function updateEnv(
-        string memory iexecResultStorageProvider_,
-        string memory iexecResultStorageProxy_
+        string memory iexecResultStorageProvider_
     ) public onlyRole(DEFAULT_ADMIN_ROLE) {
         _iexecResultStorageProvider = iexecResultStorageProvider_;
-        _iexecResultStorageProxy = iexecResultStorageProxy_;
     }
 
     /***************************************************************************
@@ -291,7 +309,11 @@ contract DataProtectorSharing is
 
         delete protectedDataDetails[_protectedData];
         collectionDetails[_collectionTokenId].size -= 1;
-        PROTECTED_DATA_REGISTRY.safeTransferFrom(address(this), msg.sender, uint256(uint160(_protectedData)));
+        PROTECTED_DATA_REGISTRY.safeTransferFrom(
+            address(this),
+            msg.sender,
+            uint256(uint160(_protectedData))
+        );
         emit ProtectedDataTransfer(_protectedData, 0, _collectionTokenId, address(0));
     }
 
@@ -327,7 +349,11 @@ contract DataProtectorSharing is
         _collectionDetails.lastSubscriptionExpiration = uint48(
             Math.max(endDate, _collectionDetails.lastSubscriptionExpiration)
         );
-        POCO_DELEGATE.transferFrom(spender, ownerOf(_collectionTokenId), _collectionDetails.subscriptionParams.price);
+        POCO_DELEGATE.transferFrom(
+            spender,
+            ownerOf(_collectionTokenId),
+            _collectionDetails.subscriptionParams.price
+        );
         emit NewSubscription(_collectionTokenId, spender, endDate);
     }
 
@@ -354,7 +380,10 @@ contract DataProtectorSharing is
     }
 
     /// @inheritdoc ISubscription
-    function setSubscriptionParams(uint256 _collectionTokenId, SubscriptionParams calldata _subscriptionParams) public {
+    function setSubscriptionParams(
+        uint256 _collectionTokenId,
+        SubscriptionParams calldata _subscriptionParams
+    ) public {
         _checkCollectionOperator(_collectionTokenId);
 
         collectionDetails[_collectionTokenId].subscriptionParams = _subscriptionParams;
@@ -405,7 +434,10 @@ contract DataProtectorSharing is
     }
 
     /// @inheritdoc IRental
-    function setProtectedDataToRenting(address _protectedData, RentingParams calldata _rentingParams) public {
+    function setProtectedDataToRenting(
+        address _protectedData,
+        RentingParams calldata _rentingParams
+    ) public {
         ProtectedDataDetails storage _protectedDataDetails = protectedDataDetails[_protectedData];
         uint256 _collectionTokenId = _protectedDataDetails.collection;
         _checkCollectionOperator(_collectionTokenId);
@@ -436,14 +468,23 @@ contract DataProtectorSharing is
         return _buyProtectedData(_protectedData, msg.sender, _to, _price);
     }
 
-    function _buyProtectedData(address _protectedData, address spender, address _to, uint72 _price) private {
+    function _buyProtectedData(
+        address _protectedData,
+        address spender,
+        address _to,
+        uint72 _price
+    ) private {
         ProtectedDataDetails storage _protectedDataDetails = protectedDataDetails[_protectedData];
         _checkProtectedDataForSale(_protectedData);
         if (_protectedDataDetails.sellingParams.price != _price) {
             revert InvalidPriceForPurchase(_protectedData, _price);
         }
 
-        PROTECTED_DATA_REGISTRY.safeTransferFrom(address(this), _to, uint256(uint160(_protectedData)));
+        PROTECTED_DATA_REGISTRY.safeTransferFrom(
+            address(this),
+            _to,
+            uint256(uint160(_protectedData))
+        );
         POCO_DELEGATE.transferFrom(
             spender,
             ownerOf(_protectedDataDetails.collection),