+++ packages/sharing-smart-contract/contracts/ManageOrders.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /******************************************************************************
- * Copyright 2024 IEXEC BLOCKCHAIN TECH                                       *
+ * Copyright 2024-2025 IEXEC BLOCKCHAIN TECH                                       *
  *                                                                            *
  * Licensed under the Apache License, Version 2.0 (the "License");            *
  * you may not use this file except in compliance with the License.           *
@@ -17,9 +17,11 @@
  ******************************************************************************/
 pragma solidity ^0.8.24;
 
-import {IExecPocoDelegate, IexecLibOrders_v5} from "./interfaces/IExecPocoDelegate.sol";
+import {IexecLibOrders_v5} from "./interfaces/IPoCo.sol";
+import {IPoCo} from "./interfaces/IPoCo.sol";
 
 /// @custom:oz-upgrades-unsafe-allow state-variable-immutable
+// eslint-disable-next-line quotes
 abstract contract ManageOrders {
     using IexecLibOrders_v5 for IexecLibOrders_v5.OrderOperationEnum;
     using IexecLibOrders_v5 for IexecLibOrders_v5.AppOrder;
@@ -31,10 +33,13 @@ abstract contract ManageOrders {
     using IexecLibOrders_v5 for IexecLibOrders_v5.RequestOrderOperation;
 
     // ---------------------ManageOrders state----------------------------------
-    IExecPocoDelegate internal immutable POCO_DELEGATE;
-    bytes32 internal constant TAG = 0x0000000000000000000000000000000000000000000000000000000000000003; // [tee,scone]
+    IPoCo internal immutable POCO_DELEGATE;
+    bytes32 internal constant TAG =
+        0x0000000000000000000000000000000000000000000000000000000000000003; // [tee,scone]
     uint256 internal constant TRUST = 0; // No replication
     string internal _iexecResultStorageProvider;
+    // [WARNING] This value is not used anymore but it must be kept
+    // to not break deployed proxies when upgrading them.
     string internal _iexecResultStorageProxy;
     uint256 private _salt;
 
@@ -42,14 +47,16 @@ abstract contract ManageOrders {
      *                        Constructor                                      *
      **************************************************************************/
     /// @custom:oz-upgrades-unsafe-allow constructor
-    constructor(IExecPocoDelegate pocoDelegate_) {
-        POCO_DELEGATE = pocoDelegate_;
+    constructor(address pocoAddress) {
+        POCO_DELEGATE = IPoCo(pocoAddress);
     }
 
     /***************************************************************************
      *                        Functions                                        *
      **************************************************************************/
-    function _createAppOrder(address _appAddress) internal view returns (IexecLibOrders_v5.AppOrder memory) {
+    function _createAppOrder(
+        address _appAddress
+    ) internal view returns (IexecLibOrders_v5.AppOrder memory) {
         //create AppOrderOperation
         return
             IexecLibOrders_v5.AppOrder({
@@ -120,28 +127,29 @@ abstract contract ManageOrders {
         uint256 _category
     ) internal returns (IexecLibOrders_v5.RequestOrder memory) {
         //create RequestOrderOperation
-        IexecLibOrders_v5.RequestOrderOperation memory requestOrderOperation = IexecLibOrders_v5.RequestOrderOperation({
-            order: IexecLibOrders_v5.RequestOrder({
-                app: _appAddress, //address
-                appmaxprice: 0, //uint256
-                dataset: _protectedData, //address
-                datasetmaxprice: 0, //uint256
-                workerpool: _workerpoolAddress, //address
-                workerpoolmaxprice: 0, //uint256
-                requester: address(this), //address
-                volume: 1, //uint256
-                tag: TAG, //bytes32
-                category: _category, //uint256
-                trust: TRUST, //uint256
-                beneficiary: msg.sender, //address
-                callback: address(0), //address
-                params: generateParams(), //string
-                salt: getSalt(), //bytes32
+        IexecLibOrders_v5.RequestOrderOperation memory requestOrderOperation = IexecLibOrders_v5
+            .RequestOrderOperation({
+                order: IexecLibOrders_v5.RequestOrder({
+                    app: _appAddress, //address
+                    appmaxprice: 0, //uint256
+                    dataset: _protectedData, //address
+                    datasetmaxprice: 0, //uint256
+                    workerpool: _workerpoolAddress, //address
+                    workerpoolmaxprice: 0, //uint256
+                    requester: address(this), //address
+                    volume: 1, //uint256
+                    tag: TAG, //bytes32
+                    category: _category, //uint256
+                    trust: TRUST, //uint256
+                    beneficiary: msg.sender, //address
+                    callback: address(0), //address
+                    params: generateParams(), //string
+                    salt: getSalt(), //bytes32
+                    sign: new bytes(0)
+                }),
+                operation: IexecLibOrders_v5.OrderOperationEnum.SIGN, //OrderOperationEnum
                 sign: new bytes(0)
-            }),
-            operation: IexecLibOrders_v5.OrderOperationEnum.SIGN, //OrderOperationEnum
-            sign: new bytes(0)
-        });
+            });
 
         // presign
         POCO_DELEGATE.manageRequestOrder(requestOrderOperation);
@@ -156,12 +164,11 @@ abstract contract ManageOrders {
     function generateParams() private view returns (string memory) {
         return
             string.concat(
-                '{"iexec_result_encryption":true',
-                ',"iexec_result_storage_provider":"',
+                // eslint-disable-next-line quotes
+                '{"iexec_result_encryption":true', // solhint-disable-line quotes
+                ',"iexec_result_storage_provider":"', // solhint-disable-line quotes
                 _iexecResultStorageProvider,
-                '","iexec_result_storage_proxy":"',
-                _iexecResultStorageProxy,
-                '"}'
+                '"}' // solhint-disable-line quotes
             );
     }
 }